[CLEANUP] Moved functionality of TextUtils into StringUtil
