+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>org.parabot</groupId>
     <artifactId>internal-api</artifactId>
-    <version>1.4.45</version>
+    <version>1.4.46</version>
 
     <packaging>jar</packaging>
